refactor(go_backend): remove legacy lowering entrypoint
