--- compiler/go_backend/parse_ast.go
@@ -1,1 +1,0 @@
-package go_backend
